requirements.txt
streamlit>=1.28.0
groq>=0.4.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
